LayoutTests/platform/wpe/fast/mediastream/MediaStream-video-element-displays-buffer-expected.txt

Tests that the stream displays captured buffers to the video element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



 === checking pixels from front camera ===
PASS mediaDevices.getUserMedia generated a stream successfully.
videos[0].srcObject = mediaStream
videos[0].play()
PASS isPixelTransparent(buffer) is true
context.drawImage(videos[0], 0, 0, 680, 360)
FAIL isPixelTransparent(buffer) should be false. Was true.
Found: {0, 0, 0, 0}
PASS isPixelBlack(buffer) is false
FAIL isPixelTransparent(buffer) should be false. Was true.
Found: {0, 0, 0, 0}
FAIL isPixelBlack(buffer) should be true. Was false.
Found: {0, 0, 0, 0}

 === checking pixels from back camera ===
PASS mediaDevices.getUserMedia generated a stream successfully.
videos[1].srcObject = mediaStream
videos[1].play()
PASS isPixelTransparent(buffer) is true
context.drawImage(videos[1], 0, 0, 680, 360)
FAIL isPixelTransparent(buffer) should be false. Was true.
Found: {0, 0, 0, 0}
PASS isPixelBlack(buffer) is false
FAIL isPixelTransparent(buffer) should be false. Was true.
Found: {0, 0, 0, 0}
FAIL isPixelGray(buffer) should be true. Was false.
Found: {0, 0, 0, 0}
PASS successfullyParsed is true

TEST COMPLETE
